Show system default layout and allow it to be copied
The system default cannot be edited but it can be copied,
which makes creating a new layout faster.

The contact_edit_options setting does allow toggling blocks on the default layout,
although it also affects the contact edit screen.
With that caveat in the confirm dialog, this extension exposes those settings
(except for the heavy-handed "CustomData" which turns off all custom blocks,
that one can remain safely on the backend settings form).